ref(widget-builder): Extract main selection components
This prepares us to add state so we can define progressive selection.
