+++ static/app/views/dashboards/widgetBuilder/components/visualize/index.tsx
@@ -13,7 +13,7 @@ import {RadioLineItem} from 'sentry/components/forms/controls/radioGroup';
 import FieldGroup from 'sentry/components/forms/fieldGroup';
 import Input from 'sentry/components/input';
 import Radio from 'sentry/components/radio';
-import {IconDelete, IconInfo} from 'sentry/icons';
+import {IconDelete} from 'sentry/icons';
 import {t, tct} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import type {SelectValue} from 'sentry/types/core';
@@ -21,9 +21,6 @@ import {defined} from 'sentry/utils';
 import {trackAnalytics} from 'sentry/utils/analytics';
 import {WidgetBuilderVersion} from 'sentry/utils/analytics/dashboardsAnalyticsEvents';
 import {
-  type AggregateParameter,
-  type AggregationKeyWithAlias,
-  type AggregationRefinement,
   classifyTagKey,
   DEPRECATED_FIELDS,
   generateFieldAsString,
@@ -35,7 +32,6 @@ import {
 import {FieldKind} from 'sentry/utils/fields';
 import {decodeScalar} from 'sentry/utils/queryString';
 import useLocationQuery from 'sentry/utils/url/useLocationQuery';
-import useApi from 'sentry/utils/useApi';
 import useCustomMeasurements from 'sentry/utils/useCustomMeasurements';
 import useOrganization from 'sentry/utils/useOrganization';
 import useTags from 'sentry/utils/useTags';
@@ -44,6 +40,10 @@ import {DisplayType, WidgetType} from 'sentry/views/dashboards/types';
 import {SectionHeader} from 'sentry/views/dashboards/widgetBuilder/components/common/sectionHeader';
 import SortableVisualizeFieldWrapper from 'sentry/views/dashboards/widgetBuilder/components/common/sortableFieldWrapper';
 import {AggregateParameterField} from 'sentry/views/dashboards/widgetBuilder/components/visualize/aggregateParameterField';
+import {
+  ColumnCompactSelect,
+  SelectRow,
+} from 'sentry/views/dashboards/widgetBuilder/components/visualize/selectRow';
 import VisualizeGhostField from 'sentry/views/dashboards/widgetBuilder/components/visualize/visualizeGhostField';
 import {useWidgetBuilderContext} from 'sentry/views/dashboards/widgetBuilder/contexts/widgetBuilderContext';
 import useDashboardWidgetSource from 'sentry/views/dashboards/widgetBuilder/hooks/useDashboardWidgetSource';
@@ -56,15 +56,7 @@ import {type FieldValue, FieldValueKind} from 'sentry/views/discover/table/types
 import {TypeBadge} from 'sentry/views/explore/components/typeBadge';
 import {useSpanTags} from 'sentry/views/explore/contexts/spanTagsContext';
 
-type AggregateFunction = [
-  AggregationKeyWithAlias,
-  string,
-  AggregationRefinement,
-  AggregationRefinement,
-];
-
-const MAX_FUNCTION_PARAMETERS = 4;
-const NONE = 'none';
+export const NONE = 'none';
 
 const NONE_AGGREGATE = {
   textValue: t('field (no aggregate)'),
@@ -99,25 +91,18 @@ function formatColumnOptions(
           ? prettifyTagKey(option.value.meta.name)
           : option.value.meta.name,
 
-      // For the spans dataset, all of the options are measurements,
-      // so we force the number badge to show
-      trailingItems:
-        dataset === WidgetType.SPANS ? (
-          <TypeBadge kind={FieldKind.MEASUREMENT} />
-        ) : (
-          renderTag(
-            option.value.kind,
-            option.value.meta.name,
-            option.value.kind !== FieldValueKind.FUNCTION &&
-              option.value.kind !== FieldValueKind.EQUATION
-              ? option.value.meta.dataType!
-              : undefined
-          )
-        ),
+      trailingItems: renderTag(
+        option.value.kind,
+        option.value.meta.name,
+        option.value.kind !== FieldValueKind.FUNCTION &&
+          option.value.kind !== FieldValueKind.EQUATION
+          ? option.value.meta.dataType!
+          : undefined
+      ),
     }));
 }
 
-function getColumnOptions(
+export function getColumnOptions(
   dataset: WidgetType,
   selectedField: QueryFieldValue,
   fieldOptions: Record<string, SelectValue<FieldValue>>,
@@ -171,26 +156,6 @@ function getColumnOptions(
   return formatColumnOptions(dataset, fieldValues, columnFilterMethod);
 }
 
-function validateParameter(
-  columnOptions: Array<SelectValue<string>>,
-  parameter: AggregateParameter,
-  value: string | undefined
-) {
-  if (parameter.kind === 'dropdown') {
-    return Boolean(parameter.options.find(option => option.value === value)?.value);
-  }
-  if (parameter.kind === 'column') {
-    return Boolean(columnOptions.find(option => option.value === value)?.value);
-  }
-  if (parameter.kind === 'value') {
-    if (parameter.dataType === 'number') {
-      return !isNaN(Number(value));
-    }
-    return true;
-  }
-  return false;
-}
-
 function canDeleteField(
   dataset: WidgetType,
   selectedFields: QueryFieldValue[],
@@ -207,15 +172,6 @@ function canDeleteField(
   return true;
 }
 
-function renderDropdownMenuFooter() {
-  return (
-    <FooterWrapper>
-      <IconInfo size="xs" />
-      {t('Select relevant fields or tags to use as groups within the table')}
-    </FooterWrapper>
-  );
-}
-
 interface VisualizeProps {
   error?: Record<string, any>;
   setError?: (error: Record<string, any>) => void;
@@ -224,7 +180,6 @@ interface VisualizeProps {
 function Visualize({error, setError}: VisualizeProps) {
   const [activeId, setActiveId] = useState<string | null>(null);
   const organization = useOrganization();
-  const api = useApi();
   const {state, dispatch} = useWidgetBuilderContext();
   let tags = useTags();
   const {customMeasurements} = useCustomMeasurements();
@@ -412,13 +367,17 @@ function Visualize({error, setError}: VisualizeProps) {
                   : field.kind === FieldValueKind.FUNCTION
                     ? datasetConfig.filterAggregateParams
                     : datasetConfig.filterTableOptions;
-                const columnOptions = getColumnOptions(
-                  state.dataset ?? WidgetType.ERRORS,
-                  field,
-                  fieldOptions,
-                  // If no column filter method is provided, show all options
-                  columnFilterMethod ?? (() => true)
-                );
+                const columnOptions =
+                  state.dataset === WidgetType.SPANS &&
+                  field.kind !== FieldValueKind.FUNCTION
+                    ? spanColumnOptions
+                    : getColumnOptions(
+                        state.dataset ?? WidgetType.ERRORS,
+                        field,
+                        fieldOptions,
+                        // If no column filter method is provided, show all options
+                        columnFilterMethod ?? (() => true)
+                      );
 
                 let aggregateOptions: Array<
                   | {
@@ -588,345 +547,22 @@ function Visualize({error, setError}: VisualizeProps) {
                             />
                           ) : (
                             <Fragment>
-                              <PrimarySelectRow hasColumnParameter={hasColumnParameter}>
-                                <AggregateCompactSelect
-                                  searchable
-                                  hasColumnParameter={hasColumnParameter}
-                                  disabled={aggregateOptions.length <= 1}
-                                  options={aggregateOptions}
-                                  value={
-                                    parseFunction(stringFields?.[index] ?? '')?.name ??
-                                    NONE
-                                  }
-                                  position="bottom-start"
-                                  menuFooter={
-                                    state.displayType === DisplayType.TABLE
-                                      ? renderDropdownMenuFooter
-                                      : undefined
-                                  }
-                                  onChange={dropdownSelection => {
-                                    const isNone = dropdownSelection.value === NONE;
-                                    let newFields = cloneDeep(fields);
-                                    const currentField = newFields[index]!;
-                                    const selectedAggregate = aggregates.find(
-                                      option =>
-                                        option.value.meta.name === dropdownSelection.value
-                                    );
-                                    // Update the current field's aggregate with the new aggregate
-                                    if (!selectedAggregate && !isNone) {
-                                      const functionFields = newFields.filter(
-                                        newField =>
-                                          newField.kind === FieldValueKind.FUNCTION
-                                      );
-                                      // Handles selection of release tags from aggregate dropdown
-                                      if (
-                                        state.dataset === WidgetType.RELEASE &&
-                                        state.displayType === DisplayType.TABLE &&
-                                        functionFields.length === 1
-                                      ) {
-                                        newFields = [
-                                          {
-                                            kind: FieldValueKind.FIELD,
-                                            field: dropdownSelection.value as string,
-                                          },
-                                          ...newFields,
-                                        ];
-
-                                        const atLeastOneFunction = newFields.some(
-                                          newField =>
-                                            newField.kind === FieldValueKind.FUNCTION
-                                        );
-
-                                        // add a function in the off chance the user gets into a state where
-                                        // they don't already have a function there
-                                        if (!atLeastOneFunction) {
-                                          newFields = [
-                                            ...newFields,
-                                            datasetConfig.defaultField,
-                                          ];
-                                        }
-                                      } else {
-                                        // Handles new selection of a field from the aggregate dropdown
-                                        newFields[index] = {
-                                          kind: FieldValueKind.FIELD,
-                                          field: dropdownSelection.value as string,
-                                        };
-                                      }
-                                      trackAnalytics(
-                                        'dashboards_views.widget_builder.change',
-                                        {
-                                          builder_version: WidgetBuilderVersion.SLIDEOUT,
-                                          field: 'visualize.updateAggregate',
-                                          from: source,
-                                          new_widget: !isEditing,
-                                          value: 'direct_column',
-                                          widget_type: state.dataset ?? '',
-                                          organization,
-                                        }
-                                      );
-                                    } else if (!isNone) {
-                                      if (currentField.kind === FieldValueKind.FUNCTION) {
-                                        // Handle setting an aggregate from an aggregate
-                                        currentField.function[0] =
-                                          dropdownSelection.value as AggregationKeyWithAlias;
-                                        if (
-                                          selectedAggregate?.value.meta &&
-                                          'parameters' in selectedAggregate.value.meta
-                                        ) {
-                                          // There are aggregates that have no parameters, so wipe out the argument
-                                          // if it's supposed to be empty
-                                          if (
-                                            selectedAggregate.value.meta.parameters
-                                              .length === 0
-                                          ) {
-                                            currentField.function[1] = '';
-                                          } else {
-                                            // Check if the column is a valid column for the new aggregate
-                                            const newColumnOptions = getColumnOptions(
-                                              state.dataset ?? WidgetType.ERRORS,
-                                              currentField,
-                                              fieldOptions,
-                                              // If no column filter method is provided, show all options
-                                              columnFilterMethod ?? (() => true)
-                                            );
-                                            const selectedAggregateIsApdexOrUserMisery =
-                                              selectedAggregate?.value.meta.name ===
-                                                'apdex' ||
-                                              selectedAggregate?.value.meta.name ===
-                                                'user_misery';
-                                            const isValidColumn =
-                                              !selectedAggregateIsApdexOrUserMisery &&
-                                              Boolean(
-                                                newColumnOptions.find(
-                                                  option =>
-                                                    option.value ===
-                                                    currentField.function[1]
-                                                )?.value
-                                              );
-                                            currentField.function[1] =
-                                              (isValidColumn
-                                                ? currentField.function[1]
-                                                : selectedAggregate.value.meta
-                                                    .parameters[0]!.defaultValue) ?? '';
-
-                                            // Set the remaining parameters for the new aggregate
-                                            for (
-                                              let i = 1; // The first parameter is the column selection
-                                              i <
-                                              selectedAggregate.value.meta.parameters
-                                                .length;
-                                              i++
-                                            ) {
-                                              // Increment by 1 to skip past the aggregate name
-                                              currentField.function[i + 1] =
-                                                selectedAggregate.value.meta.parameters[
-                                                  i
-                                                ]!.defaultValue;
-                                            }
-                                          }
-
-                                          // Wipe out the remaining parameters that are unnecessary
-                                          // This is necessary for transitioning between aggregates that have
-                                          // more parameters to ones of fewer parameters
-                                          for (
-                                            let i =
-                                              selectedAggregate.value.meta.parameters
-                                                .length;
-                                            i < MAX_FUNCTION_PARAMETERS;
-                                            i++
-                                          ) {
-                                            currentField.function[i + 1] = undefined;
-                                          }
-                                        }
-                                      } else {
-                                        if (
-                                          !selectedAggregate ||
-                                          !('parameters' in selectedAggregate.value.meta)
-                                        ) {
-                                          return;
-                                        }
-
-                                        // Handle setting an aggregate from a field
-                                        const newFunction: AggregateFunction = [
-                                          dropdownSelection.value as AggregationKeyWithAlias,
-                                          ((selectedAggregate?.value.meta?.parameters
-                                            .length > 0 &&
-                                            currentField.field) ||
-                                            selectedAggregate?.value.meta?.parameters?.[0]
-                                              ?.defaultValue) ??
-                                            '',
-                                          selectedAggregate?.value.meta?.parameters?.[1]
-                                            ?.defaultValue ?? undefined,
-                                          selectedAggregate?.value.meta?.parameters?.[2]
-                                            ?.defaultValue ?? undefined,
-                                        ];
-                                        const newColumnOptions = getColumnOptions(
-                                          state.dataset ?? WidgetType.ERRORS,
-                                          {
-                                            kind: FieldValueKind.FUNCTION,
-                                            function: newFunction,
-                                          },
-                                          fieldOptions,
-                                          // If no column filter method is provided, show all options
-                                          columnFilterMethod ?? (() => true)
-                                        );
-                                        if (
-                                          selectedAggregate?.value.meta &&
-                                          'parameters' in selectedAggregate.value.meta
-                                        ) {
-                                          selectedAggregate?.value.meta.parameters.forEach(
-                                            (parameter, parameterIndex) => {
-                                              const isValidParameter = validateParameter(
-                                                newColumnOptions,
-                                                parameter,
-                                                newFunction[parameterIndex + 1]
-                                              );
-                                              // Increment by 1 to skip past the aggregate name
-                                              newFunction[parameterIndex + 1] =
-                                                (isValidParameter
-                                                  ? newFunction[parameterIndex + 1]
-                                                  : parameter.defaultValue) ?? '';
-                                            }
-                                          );
-                                        }
-                                        newFields[index] = {
-                                          kind: FieldValueKind.FUNCTION,
-                                          function: newFunction,
-                                        };
-                                      }
-                                      trackAnalytics(
-                                        'dashboards_views.widget_builder.change',
-                                        {
-                                          builder_version: WidgetBuilderVersion.SLIDEOUT,
-                                          field: 'visualize.updateAggregate',
-                                          from: source,
-                                          new_widget: !isEditing,
-                                          value: 'aggregate',
-                                          widget_type: state.dataset ?? '',
-                                          organization,
-                                        }
-                                      );
-                                    } else {
-                                      // Handle selecting NONE so we can select just a field, e.g. for samples
-                                      // If NONE is selected, set the field to a field value
-
-                                      // When selecting NONE, the next possible columns may be different from the
-                                      // possible columns for the previous aggregate. Calculate the valid columns,
-                                      // see if the current field's function argument is in the valid columns, and if so,
-                                      // set the field to a field value. Otherwise, set the field to the first valid column.
-                                      const validColumnFields = Object.values(
-                                        datasetConfig.getTableFieldOptions?.(
-                                          organization,
-                                          tags,
-                                          customMeasurements,
-                                          api
-                                        ) ?? []
-                                      ).filter(
-                                        option =>
-                                          option.value.kind !== FieldValueKind.FUNCTION &&
-                                          (datasetConfig.filterTableOptions?.(option) ??
-                                            true)
-                                      );
-                                      const functionArgInValidColumnFields =
-                                        ('function' in currentField &&
-                                          validColumnFields.find(
-                                            option =>
-                                              option.value.meta.name ===
-                                              currentField.function[1]
-                                          )) ||
-                                        undefined;
-                                      const validColumn =
-                                        functionArgInValidColumnFields?.value.meta.name ??
-                                        validColumnFields?.[0]?.value.meta.name ??
-                                        '';
-                                      newFields[index] = {
-                                        kind: FieldValueKind.FIELD,
-                                        field: validColumn,
-                                      };
-
-                                      trackAnalytics(
-                                        'dashboards_views.widget_builder.change',
-                                        {
-                                          builder_version: WidgetBuilderVersion.SLIDEOUT,
-                                          field: 'visualize.updateAggregate',
-                                          from: source,
-                                          new_widget: !isEditing,
-                                          value: 'column',
-                                          widget_type: state.dataset ?? '',
-                                          organization,
-                                        }
-                                      );
-                                    }
-                                    dispatch({
-                                      type: updateAction,
-                                      payload: newFields,
-                                    });
-                                    setError?.({...error, queries: []});
-                                  }}
-                                  triggerProps={{
-                                    'aria-label': t('Aggregate Selection'),
-                                  }}
-                                />
-                                {hasColumnParameter && (
-                                  <ColumnCompactSelect
-                                    searchable
-                                    position="bottom-start"
-                                    menuFooter={
-                                      state.displayType === DisplayType.TABLE &&
-                                      field.kind !== FieldValueKind.FUNCTION
-                                        ? renderDropdownMenuFooter
-                                        : undefined
-                                    }
-                                    options={
-                                      state.dataset === WidgetType.SPANS &&
-                                      field.kind !== FieldValueKind.FUNCTION
-                                        ? spanColumnOptions
-                                        : columnOptions
-                                    }
-                                    value={
-                                      field.kind === FieldValueKind.FUNCTION
-                                        ? parseFunction(stringFields?.[index] ?? '')
-                                            ?.arguments[0] ?? ''
-                                        : field.field
-                                    }
-                                    onChange={newField => {
-                                      const newFields = cloneDeep(fields);
-                                      const currentField = newFields[index]!;
-                                      // Update the current field's aggregate with the new aggregate
-                                      if (currentField.kind === FieldValueKind.FUNCTION) {
-                                        currentField.function[1] =
-                                          newField.value as string;
-                                      }
-                                      if (currentField.kind === FieldValueKind.FIELD) {
-                                        currentField.field = newField.value as string;
-                                      }
-                                      dispatch({
-                                        type: updateAction,
-                                        payload: newFields,
-                                      });
-                                      setError?.({...error, queries: []});
-                                      trackAnalytics(
-                                        'dashboards_views.widget_builder.change',
-                                        {
-                                          builder_version: WidgetBuilderVersion.SLIDEOUT,
-                                          field: 'visualize.updateColumn',
-                                          from: source,
-                                          new_widget: !isEditing,
-                                          value:
-                                            currentField.kind === FieldValueKind.FIELD
-                                              ? 'column'
-                                              : 'aggregate',
-                                          widget_type: state.dataset ?? '',
-                                          organization,
-                                        }
-                                      );
-                                    }}
-                                    triggerProps={{
-                                      'aria-label': t('Column Selection'),
-                                    }}
-                                  />
-                                )}
-                              </PrimarySelectRow>
+                              <SelectRow
+                                field={field}
+                                index={index}
+                                hasColumnParameter={hasColumnParameter}
+                                columnOptions={columnOptions}
+                                aggregateOptions={aggregateOptions}
+                                stringFields={stringFields}
+                                error={error}
+                                setError={setError}
+                                fields={fields}
+                                source={source}
+                                isEditing={isEditing}
+                                fieldOptions={fieldOptions}
+                                columnFilterMethod={columnFilterMethod}
+                                aggregates={aggregates}
+                              />
                               {field.kind === FieldValueKind.FUNCTION &&
                                 parameterRefinements.length > 0 && (
                                   <ParameterRefinements>
@@ -1201,15 +837,6 @@ function renderTag(kind: FieldValueKind, label: string, dataType?: string) {
   return <BaseTag type={tagType}>{text}</BaseTag>;
 }
 
-export const ColumnCompactSelect = styled(CompactSelect)`
-  flex: 1 1 auto;
-  min-width: 0;
-
-  > button {
-    width: 100%;
-  }
-`;
-
 export const AggregateCompactSelect = styled(CompactSelect)<{
   hasColumnParameter: boolean;
 }>`
@@ -1310,14 +937,4 @@ const StyledFieldGroup = styled(FieldGroup)`
   width: 100%;
   padding: 0px;
   border-bottom: none;
-`;
-
-const FooterWrapper = styled('div')`
-  display: flex;
-  flex-direction: row;
-  gap: ${space(0.5)};
-  align-items: center;
-  justify-content: center;
-  color: ${p => p.theme.subText};
-  font-size: ${p => p.theme.fontSizeSmall};
 `;
